[LLHD][SV] Unify the ProceduralRegion trait in CIRCTSupport
Move the `ProceduralRegion` operation traits of the LLHD and SV dialects
to a common declaration within the CIRCT support library. This should
allow us to (temporarily) include procedural operations of one dialect
under procedural regions of another dialect during lowering, without
triggering an IR verification error. The behavior of the trait within
the dialects remains unchanged.
