+++ include/circt/Dialect/SV/SV.td
@@ -13,6 +13,7 @@
 #ifndef CIRCT_DIALECT_SV_SV
 #define CIRCT_DIALECT_SV_SV
 
+include "circt/Support/ProceduralRegionTrait.td"
 include "mlir/IR/AttrTypeBase.td"
 include "mlir/IR/OpBase.td"
 include "mlir/IR/OpAsmInterface.td"
@@ -24,10 +25,6 @@ include "circt/Dialect/SV/SVDialect.td"
 // Base class for the operation in this dialect.
 class SVOp<string mnemonic, list<Trait> traits = []> :
     Op<SVDialect, mnemonic, traits>;
-
-def ProceduralRegion : NativeOpTrait<"ProceduralRegion"> {
-  let cppNamespace = "::circt::sv";
-}
 
 def ProceduralOp : NativeOpTrait<"ProceduralOp"> {
   let cppNamespace = "::circt::sv";
