+++ include/circt/Support/ProceduralRegionTrait.td
@@ -1,0 +1,18 @@
+//===----------------------------------------------------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef CIRCT_SUPPORT_PROCDURALREGIONTRAIT_TD
+#define CIRCT_SUPPORT_PROCDURALREGIONTRAIT_TD
+
+include "mlir/IR/OpBase.td"
+
+def ProceduralRegion : NativeOpTrait<"ProceduralRegion"> {
+  let cppNamespace = "::circt";
+}
+
+#endif // CIRCT_SUPPORT_PROCDURALREGIONTRAIT_TD
